Fix for binding to header property when defining radium-datagrid-column.
Bindings to a `radium-datagrid-column` fail to update the UI.

For example:

```
<radium-datagrid-column header="[[xlate('NAME')]]" ...>
```

will fail to reader a header value (even though the corresponding `radium-datagrid-column` is populated with the computed `header` value).

With bindings, the `header` value is computed after the component is ready. The `_handleHeaderUpdate()` in `radium-datagrid` which listens for changes from the `header` property dispatched by `radium-datagrid-column` does successfully fire and update the property on the `radium-datagrid-column` instance. However, this has no effect because the `headerRow` template renders from the private `_headers[n].config.header` in `radium-datagrid` rather than the `radium-datagrid-column` instance's `header` property.

I tweaked `_handleHeaderUpdate()` to also update the internal `_headers[n].config.header` property when handling a `header` property update.